requirements.txt
google-cloud-compute==1.19.0
functions-framework==3.5.0